+++ Src/NugetDefense.NVD.API/Models/CvssData.cs
@@ -49,6 +49,6 @@ public class CvssDataV3
     [JsonPropertyName("availabilityImpact")]
     public string? AvailabilityImpact { get; set; }
 
-    [JsonPropertyName("baseScore")] public double BaseScore { get; set; }
+    [JsonPropertyName("baseScore")] public decimal BaseScore { get; set; }
     [JsonPropertyName("baseSeverity")] public string BaseSeverity { get; set; }
 }
